requirements.txt
streamlit
ultralytics
opencv-python
easyocr
numpy
torch
torchvision
pillow
